Get user-dependent stuff (rvm, pyenv) working
A big thanks to @harrygallagher4!
